move todoStore to different file
