+++ .eslintignore
@@ -1,2 +1,3 @@
 /build
 /node_modules
+fetchClient.ts
